remove warnings

+++ winbomb/stages.h
@@ -12,8 +12,8 @@ CHAR nd[INPUTLEN];
 
 DWORD Stage1(VOID);
 DWORD Stage2(PCHAR data);
-DWORD Stage3(DWORD data);
+DWORD Stage3(SIZE_T data);
 DWORD Stage4(PCHAR data);
-DWORD Stage5(DWORD data_len);
+DWORD Stage5(SIZE_T data_len);
 
 /* end of file */
